Dev (#59)
* Removed ImGui source

* Added ImGui submodule

* Fixed imgui breaking changes

* Removed dev branch in CI pipeline

* Workaround with filesytem

* Added Google test to submodule

* Moved "resources" directory to editor

* Adde versioning system and implement native filesystem

* Update Google test project location

* Update gitignore

* Update submodules

* Refactor

* Refactoring renderer caches

* Update imguizmo

* Refactor

* Update ImGuizmo

* Reset env map

* Added close project function and add imgui texture checking

* Refactor material & env invalidation

* Fixed material texture checking

* Fix camera resize

* Fixed material loading

* Fixed environment invalidation

* Removed test postbuild commands

+++ ignite/engine/src/ignite/graphics/objects/shadow_map.cpp
@@ -126,7 +126,7 @@ namespace ignite
 
 		glm::mat4 invView = glm::inverse(camera->GetView());
 
-		const float aspect = camera->GetViewportSize().x / camera->GetViewportSize().y;
+		const float aspect = static_cast<float>(camera->GetViewportSize().x) / static_cast<float>(camera->GetViewportSize().y);
 		const float fovRadians = glm::radians(camera->fov);
 		const float tanHalfFovY = std::tan(fovRadians * 0.5f);
 		const float tanHalfFovX = tanHalfFovY * aspect;
